requirements.txt
requests>=2.26.0
pandas==2.2.1
python-dotenv>=0.19.0
ccxt==4.2.15
web3==6.15.1
pycoingecko==3.1.0
plotly==5.18.0
aiohttp==3.9.3
asyncio==3.4.3
numpy>=1.21.0
scipy>=1.7.0
python-dateutil==2.8.2 